fix(lint): split routes/mcp_tests.rs under the 500-line gate
